Lethal Time Logit

Calculates lethal time (LT) and its fiducial confidence limits (CL) using a logit analysis according to Finney 1971, Wheeler et al. 2006, and Robertson et al. 2007.

Usage

LT_logit(formula, data, p = NULL, weights = NULL,
          subset = NULL, log_base = NULL, log_x = TRUE, het_sig = NULL,
          conf_level = NULL, long_output = TRUE)

Arguments

formula

an object of class formula or one that can be coerced to that class: a symbolic description of the model to be fitted.

data

an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which LT_logit is called.

p

Lethal time (LT) values for given p, example will return a LT50 value if p equals 50. If more than one LT value wanted specify by creating a vector. LT values can be calculated down to the 1e-16 of a percentage (e.g. LT99.99).However, the tibble produced can and will round to nearest whole number.

weights

vector of 'prior weights' to be used in the fitting process. Only needs to be supplied if you are taking the response / total for your response variable within the formula call of LC_probit. Otherwise if you use cbind(response, non-response) method you do not need to supply weights. If you do the model will be incorrect. If you don't supply weights there is a warning that will help you to make sure you are using one method or the other. See the following StackExchange post about differences cbind() function in R for a logistic regression.

subset

allows for the data to be subseted if desired. Default set to NULL.

log_base

default is 10 and will be used to calculate results using the anti of log10() given that the x variable has been log10 transformed. If FALSE results will not be back transformed.

log_x

default is TRUE and will calculate results using the antilog of determined by log_base given that the x variable has been log() transformed. If FALSE results will not be back transformed.

het_sig

significance level from person's chi square goodness-of-fit test that is used to decide if a heterogeneity factor is used. NULL is set to 0.15.

conf_level

Adjust confidence level as necessary or NULL set at 0.95.

long_output

default is TRUE which will return a tibble with all 17 variables. If FALSE the tibble returned will consist of the p level, n, the predicted LC for given p level, lower and upper confidence limits.

Value

Returns a tibble with predicted LT for given p level, lower CL (LCL), upper CL (UCL), LCL, Pearson's chi square goodness-of-fit test (pgof), slope, intercept, slope and intercept p values and standard error, and LT variance.

References

Finney, D.J., 1971. Probit Analysis, Cambridge University Press, Cambridge, England, ISBN: 052108041X

Wheeler, M.W., Park, R.M., and Bailey, A.J., 2006. Comparing median lethal concentration values using confidence interval overlap or ratio tests, Environ. Toxic. Chem. 25(5), 1441-1444.10.1897/05-320R.1

Robertson, J.L., Savin, N.E., Russell, R.M. and Preisler, H.K., 2007. Bioassays with arthropods. CRC press. ISBN: 9780849323317

Examples

head(lamprey_time)
#> # A tibble: 6 × 7
#>   nominal.dose  dose month  hour response survive total
#>          <dbl> <dbl> <chr> <int>    <int>   <int> <int>
#> 1          1.1  1.25 May       1        0      58    58
#> 2          1.1  1.25 May       2        0      58    58
#> 3          1.1  1.25 May       3        0      58    58
#> 4          1.1  1.25 May       4        1      57    58
#> 5          1.1  1.25 May       5        5      53    58
#> 6          1.1  1.25 May       6       14      44    58

results <- LT_logit((response / total) ~ log10(hour),
p = c(50, 99),
weights = total,
data = lamprey_time,
subset = c(month == "May"))

# view calculated LT50 and LT99 for seasonal
# toxicity of a piscicide, 3-trifluoromethyl-4-nitrophenol, to lamprey in 2011

results
#> # A tibble: 2 × 17
#>       p     n  time   LCL   UCL chi_square    df pgof_sig     h slope slope_se
#>   <dbl> <dbl> <dbl> <dbl> <dbl>      <dbl> <int>    <dbl> <dbl> <dbl>    <dbl>
#> 1    50    11  9.97  8.85  11.9       16.6     9   0.0551  1.85  6.77    0.734
#> 2    99    11 47.6  30.0  118.        16.6     9   0.0551  1.85  6.77    0.734
#> # ℹ 6 more variables: slope_sig <dbl>, intercept <dbl>, intercept_se <dbl>,
#> #   intercept_sig <dbl>, z <dbl>, var_m <dbl>

# dose-response curve can be plotted using 'ggplot2'