Lethal concentration data for seasonal exposure to TFM for larval sea lamprey

This dataset contains 2011 seasonal larval sea lamprey exposed to the piscicide, 3-trifluoromethyl-4-nitrophenol (TFM) toxicity data.

Usage

lamprey_tox

Format

A tibble containing 64 rows and 7 columns.

nominal_dose

nominal TFM dose (mg/L)

tank

tank assignment

month

month in which toxicity test occurred

dose

averaged measured TFM dose (mg/L) over 12 h of exposure

response

number of lamprey that responded over 12 h exposure to TFM

survive

number of lamprey that survived over 12 h exposure to TFM

total

total number of lamprey that were exposed to TFM over 12 h

Source

Hlina, B.L., Birceanu, O., Robinson, C.S., Dhiyebi, H., Wilkie, M.P. In Reivew. Seasonal Variation in the Sensitivity of Invasive Sea Lampreys to the Lampricide TFM: Importance of Energy Reserves and Temperature. North American Journal of Fisheries Management

Details

For each month, new lamprey were collected from the Deer Creek, MI, and were randomly assigned to a tank for exposure to varying doses of the piscicide, 3-trifluoromethyl-4-nitrophenol (TFM).